타 IDE 지원용 표준화 노가다 작업

+++ code/Prop/Prop_drawSelf.cpp
@@ -1,6 +1,7 @@
 
 #include <SDL3/SDL.h>
 
+import std;
 import Prop;
 import util;
 import globalVar;
@@ -14,11 +15,11 @@ import globalTime;
 import drawText;
 import ContextMenu;
 import Loot;
+import Item;
 
 
 void Prop::drawSelf()
 {
-    // RAMP prop은 prop sprite를 그리지 않음 — 화살표 마커는 renderTile.ixx::drawRampArrows에서 별도 처리
     if (leadItem.checkFlag(itemFlag::RAMP_UP) || leadItem.checkFlag(itemFlag::RAMP_DOWN)) return;
 
     constexpr Uint8 HIDE_WIRE_ALPHA = 120;
@@ -242,7 +243,6 @@ void Prop::drawSelf()
     }
 
     //롤업도어: H는 좌우 벽 유무로 변형 선택(+0 양쪽벽X / +1 우벽만 / +2 양쪽벽 / +3 좌벽만), V는 변형 없음
-    //인접 롤업도어도 벽 취급 — 여러 칸 폭의 차고문이 끊김 없이 이어져 보이게
     //열림(ROLLUP_DOOR_OPEN) 시 +16 스프라이트 + 반투명으로 말려 올라간 모습 표현
     if (leadItem.checkFlag(itemFlag::ROLLUP_DOOR))
     {
@@ -802,7 +802,7 @@ void Prop::drawSelf()
             else if (getSeason() == seasonFlag::autumn) { extraSprIndex += 2; }
             else if (getSeason() == seasonFlag::winter) { extraSprIndex += 3; }
         }
-        SDL_Point pt = { 24.0 * zoomScale,40.0 * zoomScale };
+        SDL_Point pt = { static_cast<int>(24.0 * zoomScale),static_cast<int>(40.0 * zoomScale) };
         drawSpriteCenterRotate
         (
             spr::propset,
